+++ packages/keyring-api/src/api/v2/keyring.test-d.ts
@@ -63,6 +63,7 @@ expectAssignable<KeyringCapabilities>({
   bip44: {
     derivePath: true,
     deriveIndex: true,
+    deriveIndexRange: true,
     discover: false,
   },
 });