+++ packages/keyring-eth-trezor/src/trezor-keyring-v2.test.ts
@@ -204,6 +204,7 @@ describe('TrezorKeyringV2', () => {
         bip44: {
           deriveIndex: true,
           derivePath: true,
+          deriveIndexRange: false,
           discover: false,
         },
       });